-Fixed a nasty bug where TGDS FS would be compiled and linked twice. Now it's compiled as standalone library, and linked once. -Fixed most TGDS compiler warnings, should fix random issues when linking to TGDS Projects
newlib-nds TWL:
-TGDS projects now reload its TWL ARM7 payload from 0x02800000

+++ src/common/libhardware/powerTGDS.c
@@ -48,7 +48,7 @@ void powerON(uint32 values){
 		struct sIPCSharedTGDS * TGDSIPC = getsIPCSharedTGDS();
 		uint32 * fifomsg = (uint32 *)&TGDSIPC->fifoMesaggingQueue[0];
 		setValueSafe(&fifomsg[60], (uint32)values);
-		SendFIFOWords(FIFO_POWERCNT_ON, fifomsg);
+		SendFIFOWords(FIFO_POWERCNT_ON, (u32)fifomsg);
 	}
 	else{
 		REG_POWERCNT |= values;
@@ -68,7 +68,7 @@ void powerOFF(uint32 values){
 		struct sIPCSharedTGDS * TGDSIPC = getsIPCSharedTGDS();
 		uint32 * fifomsg = (uint32 *)&TGDSIPC->fifoMesaggingQueue[0];
 		setValueSafe(&fifomsg[60], (uint32)values);
-		SendFIFOWords(FIFO_POWERCNT_OFF, fifomsg);
+		SendFIFOWords(FIFO_POWERCNT_OFF, (u32)fifomsg);
 	}
 	else{
 		REG_POWERCNT &= ~values;
